HW01_LLM/pyproject.toml
[project]
name = "hw01-llm"
version = "0.1.0"
description = "Jednoducha ukazka volani OpenAI LLM API ve smycce."
requires-python = ">=3.10"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.1",
]
